Split up lexer, tokens, parser
Split up the lexer, tokens, and parser so they're a little more
compartmentalized. For ease of maintenance, replace the parser with a
boring recursive descent parser. This is partly necessary to make it
easier to parse out nested expressions, which don't play especially
nicely with the earlier janky parser that just skimmed tokens and
parsed them as a sequence without much regard for how things were
broken up and parsed (in particular, variable access was very flaky --
and it still is, but not as much).

The lexer's been adjusted to reduce semicolons and newlines to a
single Stop token, indicating the end of a command. Continuations and
whitespace are now also treated the same, meaning that it's possible
for a Whitespace token to run into another (it would be necessary to
look ahead a rune for the lexer to run these together, which isn't
really worth the time), with one indicating the whitespace token from
before and one indicating the continuation.

Next up is to try to build a simple interpreter using the basic parsed
AST and see how that goes.