Replace string arrays with arbitrary value arrays
This is to support things like vectors, maps, strings, numbers,
and potentially other data as needed in the interpreter instead of
treating everything as a flat list of strings (with no nesting). With
this change it's possible to create just lists and strings, but lists
can be nested and so on.