Refactor for callbacks hack
Some necessary refactoring, partly for clean-up and partly to make
supporting callbacks easier.

Remove the `Context` and `Interp` types and rename `interpScope` to
`Interp`. This is to make it easier to build interpreter features
for now since the old `Interp` was never passed into anything and
the only relevant context type was `interpScope`. The `Context`
type may come back later if/when I decide to split the parser and
interpreter up to make it easier, especially if I want to move the
prelude functions into another package.

Add a Func type that carries a Cmd and an optional set of argument
bindings to pass to the Cmd before other arguments. This is produced
by both the `fn` command and the `&` (informally 'bind') command. The
latter creates a `Func` with bound arguments, similar to if you
used `fn` to wrap another function. Importantly, `&` will produce a
`Func` that survives function renames, deletions, etc. while the `fn`
wrapper relies on looking up the function each time.

Each `Interp` now carries an `overlays` slice of other `Interp`
instances that influence `upvar` and command lookups, taking precedence
over the caller `Interp` for those. Forking an `Interp` retains the
overlays slice to ensure nesting multiple layers of procs continues
working. This might be replaced with flattening overlays down to
just vars and making each `Fork` call clone the previous commands
map instead of the linked list for lookups -- this has the effect of
making function bindings sticky for each scope, though. Would either
need to require that functions are defined in order to be accessed
or retain the `Interp` traversal for functions defined out of order
from one another. For now, this works, hacky as it is.

Additional changes:

  - Add a ParseCommand function to the parser to parse just the next
    command in line. This is useful with things like rlwrap if writing
    a small REPL (which isn't included in this commit).

  - Add a Map type and dict function, along with changing `nth` to `.`
    for indexing lists and arrays. Should also probably add support
    for indexing strings but that's not here yet.

  - Rename `proc` to `fn`. Add a `-local` option to `fn` to bind the
    function to the current scope instead of global scope. This should
    probably be the default but isn't yet.

  - Move the prelude functions into prelude.go and add a Prelude
    function to return a command map for them. Add a bulk BindCmds
    function and a SetPrelude to set built-in functions.

  - Split commands into prelude and regular commands. Regular commands
    can shadow prelude commands, but if undefined will not cause the
    prelude commands to be overwritten and lost.