Fix for ints replacing strings that don't match 1:1
Fix previous behavior (that was supposed to be already committed) that
would allow an integer with, for example, a + sign to replace literal
text with an integer that did not match the original literal. May
later add something to store literal text in integers generated from
literals to keep their original string value, but not right now.